requirements.txt
# /telegram-stock-bot/requirements.txt

python-telegram-bot==21.0.1
mysql-connector-python==8.3.0
python-dotenv==1.0.1
